requirements.txt
unstructured==0.15.14
unstructured-client==0.26.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
PyPDF2==3.0.1
llama-index-vector-stores-pinecone
llama-index-core
python-dotenv
llama-index-readers-file
llama-index-vector-stores-pinecone==0.2.1
pinecone-client==5.0.1
llama-index-embeddings-openai==0.2.5
python-dotenv==1.0.1
llama-index-llms-openai==0.2.15
llama-index-readers-file==0.2.2
uvicorn==0.32.0
fastapi==0.115.2
langchain-openai==0.2.3
langchain==0.3.3
streamlit==1.39.0
llama-index-postprocessor-cohere-rerank==0.2.1
